+++ src/index.ts
@@ -2,5 +2,6 @@ export { BaseState } from "./BaseState";
 export { ActionLocator } from "./ActionLocator";
 export { StateFactory } from "./StateFactory";
 export { ModelExecutor } from "./ModelExecutor";
+export { createExecutor } from "./createExecutor";
 export { resolveStatePaths } from "./utils";
 export * from "./types";